requirements.txt
kivy
kivymd
numpy
matplotlib 
python-docx
pymongo
